Shinobi Rush V.1.0
This commit marks the initial upload of the game files for Shinobi Rush. The repository now contains the HTML, JavaScript, and CSS files necessary to run the game. Players can now access the game and enjoy navigating the ninja through challenging obstacle courses. Further updates and improvements will be made to enhance the gaming experience.